fix: clear the mechanical conversion warnings in net and backend code
Resolves -Wconversion, -Wsign-conversion and -Wimplicit-int-conversion
warnings where the conversion is known to be safe.

net_headers.h:

* ntohs() casts its result to uint16_t; the byte swap is computed in
  int.

* The UDP and TCP checksums cast the odd trailing byte and the TCP
  data-offset byte to uint32_t before shifting, so the shift and the
  addition to the checksum sum are both unsigned.

* tcp_checksum() casts the TCP length to uint16_t, the width of the
  field in the pseudo-header.

rdma_backend_tcp.c: the receive loops add recv()'s result to the byte
count as size_t. Both loops return or retry on a result of zero or
less, so the value is always positive.

rdma_backend_loopback.c: pvrdma_cqe.qp is 64 bits wide in the VMware
ABI but holds a 32-bit QP handle, so it is narrowed to uint32_t
explicitly where it is passed on.

+++ src/from-qemu/hw/rdma/rdma_backend_loopback.c
@@ -1558,8 +1558,8 @@ static void loopback_post_send(RdmaBackendDev *backend_dev, RdmaBackendQP *qp,
         remote_addr = comp_ctx->remote_addr;
         rkey = comp_ctx->rkey;
         wc_opcode = comp_ctx->cqe.opcode;
-        pvrdma_qp_handle =
-            comp_ctx->cqe.qp; /* QP handle from completion context */
+        /* QP handle from completion context */
+        pvrdma_qp_handle = (uint32_t)comp_ctx->cqe.qp;
         rdma_info_report(">>> Loopback: post_send: Extracted "
                          "pvrdma_qp_handle=%u from comp_ctx",
                          pvrdma_qp_handle);
@@ -1801,8 +1801,9 @@ static void loopback_post_send(RdmaBackendDev *backend_dev, RdmaBackendQP *qp,
                         uint32_t imm_len =
                             (transferred > 0) ? transferred : total_len;
                         pvrdma_queue_recv_imm_work_completion(
-                            pdev, imm_rctx->cq_handle, imm_rctx->cqe.qp,
-                            imm_rctx->cqe.wr_id, imm_len, lqp->qpn,
+                            pdev, imm_rctx->cq_handle,
+                            (uint32_t)imm_rctx->cqe.qp, imm_rctx->cqe.wr_id,
+                            imm_len, lqp->qpn,
                             comp_ctx ? comp_ctx->imm_data : 0);
                         rdma_info_report(
                             "Loopback: WRITE_WITH_IMM QP %u -> recv "
@@ -1810,9 +1811,9 @@ static void loopback_post_send(RdmaBackendDev *backend_dev, RdmaBackendQP *qp,
                             lqp->qpn, imm_remote->qpn,
                             comp_ctx ? comp_ctx->imm_data : 0, imm_len);
                         if (imm_rctx->cqe.qp > 0 && imm_len > 0) {
-                            loopback_update_byte_stats(backend_dev,
-                                                       imm_rctx->cqe.qp,
-                                                       imm_len, IBV_WC_RECV);
+                            loopback_update_byte_stats(
+                                backend_dev, (uint32_t)imm_rctx->cqe.qp,
+                                imm_len, IBV_WC_RECV);
                         }
                         g_free(imm_rctx);
                     }
